+++ packages/components/popover/src/popover.ts
@@ -1,6 +1,6 @@
 import type { CSSResultGroup, PropertyValues, TemplateResult } from 'lit';
 import type { PopoverPosition } from '@sl-design-system/shared';
-import { AnchorController, popoverPolyfillStyles } from '@sl-design-system/shared';
+import { AnchorController } from '@sl-design-system/shared';
 import { LitElement, html } from 'lit';
 import { property } from 'lit/decorators.js';
 import styles from './popover.scss.js';
@@ -14,6 +14,9 @@ let nextUniqueId = 0;
  * @slot default - Body content for the popover
  */
 export class Popover extends LitElement {
+  /** The default padding of the arrow. */
+  static arrowPadding = 16;
+
   /** The default offset of the popover to its anchor. */
   static offset = 12;
 
@@ -21,13 +24,18 @@ export class Popover extends LitElement {
   static override shadowRootOptions = { ...LitElement.shadowRootOptions, delegatesFocus: true };
 
   /** @private */
-  static override styles: CSSResultGroup = [popoverPolyfillStyles, styles];
+  static override styles: CSSResultGroup = styles;
 
   /** The default margin between the tooltip and the viewport. */
   static viewportMargin = 8;
 
   /** Controller for managing anchoring. */
-  #anchor = new AnchorController(this, { offset: Popover.offset, viewportMargin: Popover.viewportMargin });
+  #anchor = new AnchorController(this, {
+    arrowElement: '.arrow',
+    arrowPadding: Popover.arrowPadding,
+    offset: Popover.offset,
+    viewportMargin: Popover.viewportMargin
+  });
 
   /**
    * The position of popover relative to its anchor.